fix: 修复 event_poll_epoll.cpp 编译错误
问题：
- 缺少 <fcntl.h> 头文件导致 F_GETFL、F_SETFL、O_NONBLOCK 等宏未定义
- 整数到 size_t 的类型转换警告

修复：
- 添加 #include <fcntl.h> 头文件
- 修复 std::vector 初始化和索引的类型转换

测试结果：
- 所有 191 个测试通过

+++ packages/libfalcon/src/net/event_poll_epoll.cpp
@@ -13,6 +13,7 @@
 
 #include <sys/epoll.h>
 #include <unistd.h>
+#include <fcntl.h>
 #include <string.h>
 #include <errno.h>
 
@@ -138,7 +139,7 @@ int EPollEventPoll::poll(int timeout_ms) {
     }
 
     // 分配事件数组
-    std::vector<struct epoll_event> epoll_events(max_events_);
+    std::vector<struct epoll_event> epoll_events(static_cast<std::size_t>(max_events_));
 
     int nfds = epoll_wait(epoll_fd_, epoll_events.data(),
                          max_events_, timeout_ms);
@@ -152,7 +153,7 @@ int EPollEventPoll::poll(int timeout_ms) {
 
     // 处理就绪事件
     for (int i = 0; i < nfds; ++i) {
-        const auto& ev = epoll_events[i];
+        const auto& ev = epoll_events[static_cast<std::size_t>(i)];
         int fd = ev.data.fd;
 
         auto it = events_.find(fd);
